San Francisco COVID-19 Cases Summarized by Age Group

This dataset represents the COVID-19 positive confirmed cases by age group. Demographic and transmission data are based on information reported from case interviews, laboratories, and providers. This data may not be immediately available for recently reported cases and data will change to reflect as information becomes available. Cumulative counts of 5 or fewer are excluded from the dataset. More information about the data available here

covid19sf_age

Format

An object class data.frame with 5 variables

specimen_collection_date

date which case was recorded in YYYY-MM-DD format.

age_group

case age group c("under 18", "18-30", "31-40", "41-50", "51-60", "71-80")

new_confirmed_cases

Daily new confirmed cases

cumulative_confirmed_cases

Cumulative numbero of confirmed cases

last_updated

The table last update time in POSIX format

Source

San Francisco, Department of Public Health - Population Health Division through San Francisco Opne Data protal website.

Details

The dataset contains the daily summary of covid19 cases in San Francisco by age group

Examples

data(covid19sf_age)

head(covid19sf_age)
#>   specimen_collection_date age_group new_confirmed_cases
#> 1               2020-03-13     25-29                   3
#> 2               2020-03-14     25-29                   2
#> 3               2020-03-15     25-29                   1
#> 4               2020-03-16     25-29                   1
#> 5               2020-03-17     25-29                   0
#> 6               2020-03-18     25-29                   3
#>   cumulative_confirmed_cases        last_updated
#> 1                          7 2021-11-01 16:00:03
#> 2                          9 2021-11-01 16:00:03
#> 3                         10 2021-11-01 16:00:03
#> 4                         11 2021-11-01 16:00:03
#> 5                         11 2021-11-01 16:00:03
#> 6                         14 2021-11-01 16:00:03