San Francisco COVID-19 Cases Summarized by Date, Transmission and Case Disposition

This dataset represents the COVID-19 positive confirmed cases by race and ethnicity. Demographic data are based on information reported from case interviews, laboratories, and providers. This data may not be immediately available for recently reported cases and data will change to reflect as information becomes available. Cumulative counts of 5 or fewer are excluded from the dataset. More information about the data available here

covid19sf_demo

Format

A data frame with 5 variables.

specimen_collection_date

date which case was recorded in YYYY-MM-DD format.

race_ethnicity

The cases race/ethnicity

new_confirmed_cases

Number of daily confirmed cases

cumulative_confirmed_cases

Cumulative confirmed cases

last_updated

The table last update time in POSIX format

Source

San Francisco, Department of Public Health - Population Health Division through San Francisco Opne Data protal website.

Details

The dataset contains the daily summary of covid19 cases in San Francisco by race/ethnicity group

Examples

data(covid19sf_demo)

head(covid19sf_demo)

#>   specimen_collection_date race_ethnicity new_confirmed_cases
#> 1               2020-04-18          Asian                   0
#> 2               2020-05-18          Asian                   4
#> 3               2020-05-30          Asian                   3
#> 4               2020-05-31          Asian                   5
#> 5               2020-06-01          Asian                   5
#> 6               2020-06-02          Asian                   3
#>   cumulative_confirmed_cases        last_updated
#> 1                        198 2020-09-06 15:19:27
#> 2                        305 2020-09-06 15:19:27
#> 3                        335 2020-09-06 15:19:27
#> 4                        340 2020-09-06 15:19:27
#> 5                        345 2020-09-06 15:19:27
#> 6                        348 2020-09-06 15:19:27