San Francisco COVID-19 Cases Summarized by Homelessness Status

This dataset represents the COVID-19 positive confirmed cases by homelessness. Demographic and transmission data are based on information reported from case interviews, laboratories, and providers. This data may not be immediately available for recently reported cases and data will change to reflect as information becomes available. Cumulative counts of 5 or fewer are excluded from the dataset. More information about the data available here

covid19sf_homeless

Format

An object class data.frame with 5 variables

specimen_collection_date

Date which case was recorded in YYYY-MM-DD format.

homelessness_status

The homelessness status, a signle category variable c("Homeless")

new_confirmed_cases

Total cases confirmed cases per date

cumulative_confirmed_cases

Cumulative confirmed cases

last_updated

The table last update time in POSIX format

Source

San Francisco, Department of Public Health - Population Health Division through San Francisco Opne Data protal website.

Details

The dataset contains the daily summary of covid19 cases in San Francisco by homelessness status

Examples

data(covid19sf_homeless)

head(covid19sf_homeless)

#>   specimen_collection_date homelessness_status new_confirmed_cases
#> 1               2020-03-31            Homeless                   3
#> 2               2020-04-01            Homeless                   0
#> 3               2020-04-02            Homeless                   1
#> 4               2020-04-03            Homeless                   0
#> 5               2020-04-04            Homeless                   1
#> 6               2020-04-05            Homeless                   1
#>   cumulative_confirmed_cases        last_updated
#> 1                          5 2020-09-26 14:47:40
#> 2                          5 2020-09-26 14:47:40
#> 3                          6 2020-09-26 14:47:40
#> 4                          6 2020-09-26 14:47:40
#> 5                          7 2020-09-26 14:47:40
#> 6                          8 2020-09-26 14:47:40