San Francisco COVID-19 Hospitalizations

Count of COVID+ patients admitted to the hospital. Patients who are hospitalized and test positive for COVID-19 may be admitted to an acute care bed (a regular hospital bed), or an intensive care unit (ICU) bed. This data shows the daily total count of COVID+ patients in these two bed types, and the data reflects totals from all San Francisco Hospitals. More information about the data available here

covid19sf_hospitalizations

Format

An object class data.frame with 5 variables

reportdate

date which case was recorded in YYYY-MM-DD format.

hospital

The hospital which patients were admitted, currently it labeled under "All SF Hospitals"

dphcategory

The type of hospitalization bed, either an acute care bed (a regular hospital bed), or an intensive care unit (ICU) bed

covidstatus

The patient diagnostic, either PUI (Patient Under Investigation) or COVID+ (positive case)

patientcount

Daily cases count

Source

San Francisco, Department of Public Health - Population Health Division through San Francisco Opne Data protal website.

Details

Each record represents how many people were hospitalized on the date recorded in either an ICU bed or acute care bed (shown as Med/Surg under DPHCategory field)

Examples

data(covid19sf_hospitalizations)

head(covid19sf_hospitalizations)
#>   reportdate         hospital dphcategory covidstatus patientcount
#> 1 2020-03-23 All SF Hospitals    Med/Surg      COVID+           12
#> 2 2020-03-23 All SF Hospitals    Med/Surg         PUI          101
#> 3 2020-03-24 All SF Hospitals    Med/Surg      COVID+           20
#> 4 2020-03-24 All SF Hospitals    Med/Surg         PUI           96
#> 5 2020-03-25 All SF Hospitals    Med/Surg      COVID+           22
#> 6 2020-03-25 All SF Hospitals    Med/Surg         PUI           62
#>            data_loaded_at
#> 1 2021-12-17T09:00:09.024
#> 2 2021-12-17T09:00:11.360
#> 3 2021-12-17T09:00:11.361
#> 4 2021-12-17T09:00:11.363
#> 5 2021-12-17T09:00:11.364
#> 6 2021-12-17T09:00:11.366