San Francisco COVID-19 Alternative Housing Sites

This dataset includes aggregate data on the type, status, population served, and individuals placed at each alternative housing site under contract with HSA. More information about the data available here

covid19sf_housing

Format

A data frame with 8 variables.

site_id

Site ID

status

The site status, c("Active", "In Preparation")

facility_type

The facility type, c("Hotel", "Safe Sleep", "Congregate", "RV")

site_type

The site type, c(""SIP: COVID-Negative/Unknown", "I/Q", "SS: COVID-Negative/Unknown", "SIP: Post-COVID")

units_occupied

Number of units occupied per site

total_units

Total number of units available

population_covid_status

The population covid status, c("COVID Negative/Unknown", "COVID Positive", "Post-COVID")

date_updated

Date which data was updated in YYYY-MM-DD format.)

Source

San Francisco, Department of Public Health - Population Health Division through San Francisco Opne Data protal website.

Details

The dataset contains a summary of covid19 housing site in San Francisco by site, facility and covid19 status

Examples

data(covid19sf_housing)

head(covid19sf_housing)

#>   site_id status facility_type                   site_type units_occupied
#> 1     1.0 Active         Hotel SIP: COVID-Negative/Unknown             44
#> 2     2.0 Active         Hotel SIP: COVID-Negative/Unknown            131
#> 3     4.0 Active         Hotel SIP: COVID-Negative/Unknown            107
#> 4     5.0 Active         Hotel SIP: COVID-Negative/Unknown             58
#> 5     6.0 Active         Hotel SIP: COVID-Negative/Unknown             26
#> 6     7.0 Active         Hotel SIP: COVID-Negative/Unknown             59
#>   total_units population_covid_status date_updated
#> 1          59  COVID Negative/Unknown   2020-09-11
#> 2         142  COVID Negative/Unknown   2020-09-11
#> 3         116  COVID Negative/Unknown   2020-09-11
#> 4          63  COVID Negative/Unknown   2020-09-11
#> 5          31  COVID Negative/Unknown   2020-09-11
#> 6          68  COVID Negative/Unknown   2020-09-11