COVID-19 Cases by Population Characteristics Over Time

TThis dataset shows San Francisco COVID-19 cases by population characteristics and by specimen collection date. Cases are included on the date the positive test was collected. Population characteristics are subgroups, or demographic cross-sections, like age, race, or gender. The City tracks how cases have been distributed among different subgroups. This information can reveal trends and disparities among groups. Data is lagged by five days, meaning the most recent specimen collection date included is 5 days prior to today. Tests take time to process and report, so more recent data is less reliable. More details available here

covid19sf_population

Format

An object class data.frame with 7 variables

specimen_collection_date

Date which case was recorded in YYYY-MM-DD format.

characteristic_type

Overall topic area for a given population characteristic. These are subgroups or demographic cross-sections, like age

characteristic_group

Each group or category within a characteristic type or topic area. ex 0-4 yrs, 5-10 yrs

characteristic_group_sort_order

Sort order of characteristic group to aid in visualizing data

new_cases

Cases are counted as confirmed on the date of specimen collection after a positive lab test result

cumulative_cases

Cumulative Cases

population_estimate

Population estimate for a given characteristic type and characteristic group

Source

San Francisco, Department of Public Health - Population Health Division through San Francisco Opne Data protal website.

Details

The dataset contains a summary of COVID-19 cases overtime by population characteristics

Examples

data(covid19sf_population)

head(covid19sf_population)
#>   specimen_collection_date characteristic_type characteristic_group
#> 1               2020-03-03           Age Group                  0-4
#> 2               2020-03-03           Age Group                 5-11
#> 3               2020-03-03           Age Group                12-17
#> 4               2020-03-03           Age Group                18-20
#> 5               2020-03-03           Age Group                21-24
#> 6               2020-03-03           Age Group                25-29
#>   characteristic_group_sort_order new_cases cumulative_cases
#> 1                               1        NA               NA
#> 2                               2        NA               NA
#> 3                               3        NA               NA
#> 4                               4        NA               NA
#> 5                               5        NA               NA
#> 6                               6        NA               NA
#>   population_estimate
#> 1               39353
#> 2               44153
#> 3               34664
#> 4               20407
#> 5               39944
#> 6              100792