San Francisco COVID-19 Testing Locations

A list of testing locations including address and coordinates for mapping. More information about the data available here

covid19sf_test_loc

Format

An object class sf and data.frame with 17 variables

id

Location ID

medical_home

Medical home

name

The medical name

address

The medical address

phone_number

The medical phone number

phone_number_formatted

The medical phone number formatted

testing_hours

The medical testing hours

popup_or_permanent

The medical testing type, c("Permanent", "Pop-Up" )

location_type

The medical location type, c("Private", "Public" )

eligibility

Eligibility information for accessing testing at this location

cta_text

The call to action used for the web map

cta_link

The call to action link for the button on the web map

sample_collection_method

The method for collecting samples at the lab

lap

The lab name

latitude

The medical latitude point

longitude

The medical longitude point

geometry

The medical geometry details

Source

San Francisco, Department of Public Health - Population Health Division through San Francisco Opne Data protal website.

Details

The dataset contains the San Francisco testing location information

Examples

data(covid19sf_test_loc)

head(covid19sf_test_loc)

#> Simple feature collection with 6 features and 16 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -122.4924 ymin: 37.71886 xmax: -122.3874 ymax: 37.79248
#> CRS:            4326
#>   id                                             medical_home
#> 1 26                                                      CVS
#> 2  1                                            Carbon Health
#> 3  7                                           Dignity Health
#> 4 12                                              One Medical
#> 5 45 SF Department of Public Health in partnership with Color
#> 6 27                                      Primary Care Clinic
#>                                name                                address
#> 1                 CVS Minute Clinic                           377 32nd Ave
#> 2     Carbon Health - Castro Clinic                         1998 Market St
#> 3     Go Health Urgent Care Centers 288 Market Street (Multiple Locations)
#> 4 One Medical Clinic - Hayes Valley                            98 Gough St
#> 5         Alice Griffith Apartments             2600 Arelious Walker Drive
#> 6  Circle Medical - Downtown Clinic                    333 1st St, Suite A
#>   phone_number phone_number_formatted                 testing_hours
#> 1         <NA>                   <NA>                          <NA>
#> 2   4157926040         (415) 792-6040                       9am-7pm
#> 3         <NA>                   <NA> 8am-8pm M-F, 9am-5pm Weekends
#> 4   4155931134         (415) 593-1134                   8am-6pm M-F
#> 5          311             Dial 3-1-1     Saturday (6/19), 10am-2pm
#> 6   4158400560         (415) 840-0560 8am-5pm M-F, 8am-2pm Weekends
#>   popup_or_permanent location_type
#> 1          Permanent        Public
#> 2          Permanent       Private
#> 3          Permanent       Private
#> 4          Permanent       Private
#> 5             Pop-Up        Public
#> 6          Permanent       Private
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        eligibility
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             Must meet CDC guidelines for Testing
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    Must be Carbon Health patient
#> 3 Virtual Visit required first for screening. Virtual Visits have zero cost to patients under Medicare/Medicaid and most other major insurance plans during the public health emergency. Currently, Virtual Visits are covered by Aetna, Anthem, Cigna, Health Plan San Mateo, Health Net (during emergency declarations) Humana, and UnitedHealthcare. Otherwise $140 out of pocket for virtual visit, which goes toward copay if visit to physical location is necessary after virtual screening, visit must happen within 24 hours of referral.
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      Must be One Medical patient
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           Anyone living or working in SF. No insurance required.
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 None. All major insurance Payers
#>                    cta_text
#> 1           Schedule a test
#> 2     Sign up with provider
#> 3     Sign up with provider
#> 4     Sign up with provider
#> 5                  Drop in.
#> 6 Schedule via phone or web
#>                                                                                                                                                                                                                               cta_link
#> 1                                                                                                                                                                                     http://www.cvs.com/minuteclinic/covid-19-testing
#> 2 https://patient.carbonhealth.com/#/schedule?practiceId=5bdaef44-8ff0-439f-99d7-3285afcc6911&vstId=0f02af4e-1d1c-4a4b-a0f2-e67fca98dd77&vstrId=f743d1d1-d2a1-4fce-a5f9-2e29339b2406&specialtyIds=652c64cc-9fcb-442f-b71e-266390ef2f63
#> 3                                                                                                                                                                                    https://www.gohealthuc.com/bayarea/virtual-visits
#> 4                                                                                                                                                               https://www.onemedical.com/blog/live-well/covid-19-testing-one-medical
#> 5                                                                                                                                                                                                           https://sf.gov/GetTestedSF
#> 6                                                                                                                                                                                                        https://www.circlemedical.com
#>   sample_collection_method                       lab   latitude    longitude
#> 1                       AN                     Quest       <NA>         <NA>
#> 2                     <NA> Rapid test (Abbot ID NOW)       <NA>         <NA>
#> 3                     <NA> Rapid test (Abbot ID NOW)       <NA>         <NA>
#> 4                     <NA>                     Quest       <NA>         <NA>
#> 5                       AN                      <NA> 37.7188588 -122.3874251
#> 6                     <NA>                      <NA>       <NA>         <NA>
#>                     geometry
#> 1 POINT (-122.4924 37.78198)
#> 2 POINT (-122.4259 37.77001)
#> 3 POINT (-122.3973 37.79248)
#> 4 POINT (-122.4224 37.77408)
#> 5 POINT (-122.3874 37.71886)
#> 6 POINT (-122.3939 37.78684)