COVID-19 Vaccine Doses Given to San Franciscans by Demographics Over Time

This dataset represents doses of COVID-19 vaccine administered in California to San Francisco residents over time. The data is broken down by multiple demographic slices. The three dose types are counted separately, i.e. (1) first doses administered as a part of a two-dose vaccination, (2) second doses administered as part of a two-dose vaccination, and (3) single-dose vaccines administered. here

covid19sf_vaccine_demo_ts

Format

An object class data.frame with 21 variables

date_administered

Date vaccination administered

overall_segment

Segment (universe) of analysis. Unique combination of administering_provider_type, age_group, and demographic_group. Filter to a single option to derive meaningful totals.

administering_provider_type

Providers included in a given overall_segment. Two possible values: 'All' (including SF DPH) or 'DPH Only'

age_group

Age range included in a given overall_segment

demographic_group

Type of demographic group included in a given overall_segment (e.g. Age, Race/Ethnicity)

demographic_subgroup

Specific demographic group counted in a given record (e.g. 16-24, Asian)

demographic_subgroup_sort_order

Numeric sort order for all demographic_subgroup. Convenient for maintaining consistent ordering across multiple data visualizations.

new_1st_doses

Count of 1st doses administered for vaccines that take two doses to complete

new_2nd_doses

Count of 2nd doses administered for vaccines that take two doses to complete

new_single_doses

Count of doses administered for vaccines that take one dose to complete

new_series_completed

Count of individuals newly fully vaccinated on a given day (given the 2nd dose of a two-dose vaccine or one dose of a single dose vaccine)

new_recipients

Count of individuals vaccinated (with any dose) for the first time according to CA's records

cumulative_1st_doses

Cumulative total of 1st doses administered for vaccines that take two doses to complete

cumulative_2nd_doses

Cumulative total of 2nd doses administered for vaccines that take two doses to complete

cumulative_single_doses

Cumulative total of doses administered for vaccines that take one dose to complete

cumulative_series_completed

Cumulative total individuals fully vaccinated (given the 2nd dose of a two-dose vaccine or one dose of a single dose vaccine)

cumulative_recipients

Cumulative total individuals vaccinated (with any dose) according to CA's records

subgroup_population

American Community Survey population estimates for given demographic_subgroup

age_group_population

American Community Survey population estimates for overall age_group

data_as_of

Timestamp for last update date in source system

data_loaded_at

Timestamp when the record (row) was most recently updated here in the Open Data Portal

Source

San Francisco, Department of Public Health - Population Health Division through San Francisco Opne Data protal website.

Details

The dataset contains a time series of COVID-19 vaccine doses given to San Franciscans by demographics

Examples

data(covid19sf_vaccine_demo_ts)

head(covid19sf_vaccine_demo_ts)
#>   date_administered                                       overall_segment
#> 1        2021-07-28 Ages 5+ by Age Bracket, Administered by All Providers
#> 2        2021-07-29 Ages 5+ by Age Bracket, Administered by All Providers
#> 3        2021-07-30 Ages 5+ by Age Bracket, Administered by All Providers
#> 4        2021-07-31 Ages 5+ by Age Bracket, Administered by All Providers
#> 5        2021-08-01 Ages 5+ by Age Bracket, Administered by All Providers
#> 6        2021-08-02 Ages 5+ by Age Bracket, Administered by All Providers
#>   administering_provider_type age_group demographic_group demographic_subgroup
#> 1               All Providers        5+       Age Bracket                12-17
#> 2               All Providers        5+       Age Bracket                12-17
#> 3               All Providers        5+       Age Bracket                12-17
#> 4               All Providers        5+       Age Bracket                12-17
#> 5               All Providers        5+       Age Bracket                12-17
#> 6               All Providers        5+       Age Bracket                12-17
#>   demographic_subgroup_sort_order new_1st_doses new_2nd_doses new_single_doses
#> 1                               3           112            95                0
#> 2                               3           123           100                0
#> 3                               3           110            83                0
#> 4                               3            77            63                0
#> 5                               3            38            18                0
#> 6                               3           112           102                0
#>   new_series_completed new_recipients cumulative_1st_doses cumulative_2nd_doses
#> 1                   95            112                31191                27942
#> 2                  100            123                31314                28042
#> 3                   83            110                31424                28125
#> 4                   63             77                31501                28188
#> 5                   18             38                31539                28206
#> 6                  102            112                31651                28308
#>   cumulative_single_doses cumulative_series_completed cumulative_recipients
#> 1                      12                       27954                 31220
#> 2                      12                       28054                 31343
#> 3                      12                       28137                 31453
#> 4                      12                       28200                 31530
#> 5                      12                       28218                 31568
#> 6                      12                       28320                 31680
#>   subgroup_population age_group_population          data_as_of
#> 1               33938               835137 2021-12-15 06:00:11
#> 2               33938               835137 2021-12-15 06:00:11
#> 3               33938               835137 2021-12-15 06:00:11
#> 4               33938               835137 2021-12-15 06:00:11
#> 5               33938               835137 2021-12-15 06:00:11
#> 6               33938               835137 2021-12-15 06:00:11
#>        data_loaded_at
#> 1 2021-12-15 08:15:07
#> 2 2021-12-15 08:15:07
#> 3 2021-12-15 08:15:07
#> 4 2021-12-15 08:15:07
#> 5 2021-12-15 08:15:07
#> 6 2021-12-15 08:15:07