San Francisco COVID-19 Alternative Housing Sites

This dataset includes aggregate data on the type, status, population served, and individuals placed at each alternative housing site under contract with HSA. More information about the data available here

covid19sf_housing

Format

An object class data.frame with 8 variables

site_id

Site ID

status

The site status, c("Active", "In Preparation")

facility_type

The facility type, c("Hotel", "Safe Sleep", "Congregate", "RV")

site_type

The site type, c(""SIP: COVID-Negative/Unknown", "I/Q", "SS: COVID-Negative/Unknown", "SIP: Post-COVID")

units_occupied

Number of units occupied per site

total_units

Total number of units available

population_covid_status

The population covid status, c("COVID Negative/Unknown", "COVID Positive", "Post-COVID")

date_updated

Date which data was updated in YYYY-MM-DD format)

Source

San Francisco, Department of Public Health - Population Health Division through San Francisco Opne Data protal website.

Details

The dataset contains a summary of covid19 housing site in San Francisco by site, facility and covid19 status

Examples

data(covid19sf_housing)

head(covid19sf_housing)
#>   site_id status facility_type                   site_type units_occupied
#> 1     1.0 Active         Hotel SIP: COVID-Negative/Unknown             20
#> 2     2.0 Active         Hotel SIP: COVID-Negative/Unknown            128
#> 3     4.0 Active         Hotel SIP: COVID-Negative/Unknown             84
#> 4     5.0 Active         Hotel SIP: COVID-Negative/Unknown             47
#> 5     7.0 Active         Hotel SIP: COVID-Negative/Unknown              2
#> 6     8.0 Active         Hotel                         I/Q            388
#>   total_units population_covid_status date_updated
#> 1          62  COVID Negative/Unknown   2021-04-26
#> 2         144  COVID Negative/Unknown   2021-04-26
#> 3         117  COVID Negative/Unknown   2021-04-26
#> 4          63  COVID Negative/Unknown   2021-04-26
#> 5          69  COVID Negative/Unknown   2021-04-26
#> 6          77          COVID Positive   2021-04-26