San Francisco COVID-19 Cases Summarized by Date, Transmission and Case Disposition

This dataset represents the COVID-19 positive confirmed cases and deaths by day and transmission type. The transmission data are based on information reported from case interviews, laboratories, and providers. This data may not be immediately available for recently reported cases and data will change to reflect as information becomes available. More information about the data available here

covid19sf_summary

Format

An object class data.frame with 5 variables

specimen_collection_date

Date which case was recorded in YYYY-MM-DD format.

case_disposition

The case disposition c("Confirmed", "Death)

transmission_category

The case transmission category c("Community", "From Contact", "Unknown")

case_count

Daily cases count

last_updated

The table last update time in POSIX format

Source

San Francisco, Department of Public Health - Population Health Division through San Francisco Opne Data protal website.

Details

The dataset contains the daily summary of covid19 cases in San Francisco by transmission and case disposition

Examples

data(covid19sf_summary)

head(covid19sf_summary)

#>   specimen_collection_date case_disposition transmission_category case_count
#> 1               2020-03-03        Confirmed             Community          2
#> 2               2020-03-05        Confirmed             Community          1
#> 3               2020-03-05        Confirmed          From Contact          5
#> 4               2020-03-07        Confirmed          From Contact          5
#> 5               2020-03-08        Confirmed          From Contact          1
#> 6               2020-03-10        Confirmed          From Contact          2
#>          last_updated
#> 1 2020-09-12 14:19:25
#> 2 2020-09-12 14:19:25
#> 3 2020-09-12 14:19:25
#> 4 2020-09-12 14:19:25
#> 5 2020-09-12 14:19:25
#> 6 2020-09-12 14:19:25

# Case disposition summary
table(covid19sf_summary$case_disposition)

#> 
#> Confirmed     Death 
#>       472        80 

# Transmission category
table(covid19sf_summary$transmission_category)

#> 
#>    Community From Contact      Unknown 
#>          232          217          103 

# Summary of case disposition and transmission category
table(covid19sf_summary$case_disposition,
      covid19sf_summary$transmission_category)

#>            
#>             Community From Contact Unknown
#>   Confirmed       187          186      99
#>   Death            45           31       4