San Francisco COVID-19 Testing Locations

A list of testing locations including address and coordinates for mapping. More information about the data available here

covid19sf_test_loc

Format

An object class sf and data.frame with 17 variables

id

Location ID

medical_home

Medical home

name

The medical name

address

The medical address

phone_number

The medical phone number

phone_number_formatted

The medical phone number formatted

testing_hours

The medical testing hours

popup_or_permanent

The medical testing type, c("Permanent", "Pop-Up" )

location_type

The medical location type, c("Private", "Public" )

eligibility

Eligibility information for accessing testing at this location

cta_text

The call to action used for the web map

cta_link

The call to action link for the button on the web map

sample_collection_method

The method for collecting samples at the lab

lap

The lab name

latitude

The medical latitude point

longitude

The medical longitude point

geometry

The medical geometry details

Source

San Francisco, Department of Public Health - Population Health Division through San Francisco Opne Data protal website.

Details

The dataset contains the San Francisco testing location information

Examples

data(covid19sf_test_loc)

head(covid19sf_test_loc)

#> Simple feature collection with 6 features and 16 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -122.4794 ymin: 37.75437 xmax: -122.4034 ymax: 37.79014
#> geographic CRS: WGS 84
#>   id                      medical_home                               name
#> 1  1                     Carbon Health      Carbon Health - Castro Clinic
#> 2  6                           Forward                     Forward Clinic
#> 3 11        Northeast Medical Services                     Noriega Clinic
#> 4 25                              MNHC Mission Neighborhood Health Center
#> 5 26 Sutter Pacific Medical Foundation  Sutter Pacific Medical Foundation
#> 6 15                              UCSF      Pediatrics Clinic at Mt. Zion
#>                         address        phone_number phone_number_formatted
#> 1                1998 Market St          4157926040         (415) 792-6040
#> 2                 180 Sutter St          8333346393         (833) 334-6393
#> 3                  1450 Noriega          4153919686         (415) 391-9686
#> 4               240 Shotwell St 4155523870 ext 2217    4155523870 ext 2217
#> 5 3801 Sacramento St, 3rd Floor          4156006204         (415) 600-6204
#> 6             2330 Post St #320          4158857478         (415) 885-7478
#>   testing_hours popup_or_permanent location_type
#> 1       9am-7pm          Permanent       Private
#> 2   8am-6pm M-F          Permanent       Private
#> 3   2pm-3pm M-F          Permanent       Private
#> 4   8am-5pm M-F             Pop-Up        Public
#> 5   9am-6pm M-F          Permanent       Private
#> 6   8am-5pm M-F          Permanent       Private
#>                                                         eligibility
#> 1                                     Must be Carbon Health patient
#> 2                                    Must be Forward Health patient
#> 3 Must be NEMS patient. May register as a new NEMS patient for free
#> 4    Anyone living or working in SF. Convenient for people on foot.
#> 5                                        Call to confirm eligibilty
#> 6                                              Must be UCSF patient
#>                      cta_text
#> 1       Sign up with provider
#> 2       Sign up with provider
#> 3       Sign up with provider
#> 4           Call to book test
#> 5           Call to book test
#> 6 Physician referral required
#>                                                                                                                                                                                                                               cta_link
#> 1 https://patient.carbonhealth.com/#/schedule?practiceId=5bdaef44-8ff0-439f-99d7-3285afcc6911&vstId=0f02af4e-1d1c-4a4b-a0f2-e67fca98dd77&vstrId=f743d1d1-d2a1-4fce-a5f9-2e29339b2406&specialtyIds=652c64cc-9fcb-442f-b71e-266390ef2f63
#> 2                                                                                                                                                                                                    https://goforward.com/coronavirus
#> 3                                                                                                                                                                                                        https://www.nems.org/covid19/
#> 4                                                                                                                                                                                                                     tel://4155523870
#> 5                                                                                                                                                                                                                     tel://4156006204
#> 6                                                                                                                                                                                                                                 <NA>
#>   sample_collection_method                       lab latitude longitude
#> 1                     <NA> Rapid test (Abbot ID NOW)     <NA>      <NA>
#> 2                     <NA>                      <NA>     <NA>      <NA>
#> 3                     <NA>                      <NA>     <NA>      <NA>
#> 4                     <NA>                      <NA>     <NA>      <NA>
#> 5                     <NA>                    Sutter     <NA>      <NA>
#> 6                     <NA>                      UCSF     <NA>      <NA>
#>                     geometry
#> 1 POINT (-122.4263 37.76988)
#> 2 POINT (-122.4034 37.79014)
#> 3 POINT (-122.4794 37.75437)
#> 4 POINT (-122.4165 37.76459)
#> 5 POINT (-122.4554 37.78683)
#> 6 POINT (-122.4404 37.78454)

library(mapview)

mapview(covid19sf_test_loc, zcol = "name", legend = FALSE)