San Francisco COVID-19 Testing Locations

A list of testing locations including address and coordinates for mapping. More information about the data available here

covid19sf_test_loc

Format

An object class sf and data.frame with 17 variables

id

Location ID

medical_home

Medical home

name

The medical name

address

The medical address

phone_number

The medical phone number

phone_number_formatted

The medical phone number formatted

testing_hours

The medical testing hours

popup_or_permanent

The medical testing type, c("Permanent", "Pop-Up" )

location_type

The medical location type, c("Private", "Public" )

eligibility

Eligibility information for accessing testing at this location

cta_text

The call to action used for the web map

cta_link

The call to action link for the button on the web map

sample_collection_method

The method for collecting samples at the lab

lap

The lab name

latitude

The medical latitude point

longitude

The medical longitude point

geometry

The medical geometry details

Source

San Francisco, Department of Public Health - Population Health Division through San Francisco Opne Data protal website.

Details

The dataset contains the San Francisco testing location information

Examples

data(covid19sf_test_loc)

head(covid19sf_test_loc)

#> Simple feature collection with 6 features and 16 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -122.4924 ymin: 37.76191 xmax: -122.4098 ymax: 37.78699
#> CRS:            4326
#>   id                                             medical_home
#> 1  1                                            Carbon Health
#> 2  6                                                   Kaiser
#> 3 11                                              One Medical
#> 4 44 SF Department of Public Health in partnership with Color
#> 5 26                                                      CVS
#> 6 39 SF Department of Public Health in partnership with Color
#>                              name                   address phone_number
#> 1   Carbon Health - Castro Clinic            1998 Market St   4157926040
#> 2                   French Campus           4131 Geary Blvd   4158332000
#> 3 One Medical Clinic - Drive Thru        3333 California St   4155931134
#> 4             Mission 18th & Capp 18th Street & Capp Street          311
#> 5               CVS Minute Clinic              377 32nd Ave         <NA>
#> 6                            SOMA        967 Mission Street          311
#>   phone_number_formatted                                  testing_hours
#> 1         (415) 792-6040                                        9am-7pm
#> 2         (415) 833-2000                                    8am-5pm M-F
#> 3         (415) 593-1134                                    8am-6pm M-F
#> 4             Dial 3-1-1  Saturday (6/12) and Saturday (6/26), 10am-3pm
#> 5                   <NA>                                           <NA>
#> 6             Dial 3-1-1 Wednesday (6/2) and Wednesday (6/16), 10am-4pm
#>   popup_or_permanent location_type
#> 1          Permanent       Private
#> 2          Permanent       Private
#> 3             Pop-Up       Private
#> 4             Pop-Up        Public
#> 5          Permanent        Public
#> 6             Pop-Up        Public
#>                                              eligibility              cta_text
#> 1                          Must be Carbon Health patient Sign up with provider
#> 2                               Must be a Kaiser patient Sign up with provider
#> 3                            Must be One Medical patient Sign up with provider
#> 4 Anyone living or working in SF. No insurance required.              Drop in.
#> 5                   Must meet CDC guidelines for Testing       Schedule a test
#> 6 Anyone living or working in SF. No insurance required.              Drop in.
#>                                                                                                                                                                                                                               cta_link
#> 1 https://patient.carbonhealth.com/#/schedule?practiceId=5bdaef44-8ff0-439f-99d7-3285afcc6911&vstId=0f02af4e-1d1c-4a4b-a0f2-e67fca98dd77&vstrId=f743d1d1-d2a1-4fce-a5f9-2e29339b2406&specialtyIds=652c64cc-9fcb-442f-b71e-266390ef2f63
#> 2                                                                                                                            https://healthy.kaiserpermanente.org/northern-california/health-wellness/coronavirus-information/get-care
#> 3                                                                                                                                                               https://www.onemedical.com/blog/live-well/covid-19-testing-one-medical
#> 4                                                                                                                                                                                                           https://sf.gov/GetTestedSF
#> 5                                                                                                                                                                                     http://www.cvs.com/minuteclinic/covid-19-testing
#> 6                                                                                                                                                                                                           https://sf.gov/GetTestedSF
#>   sample_collection_method                       lab   latitude    longitude
#> 1                     <NA> Rapid test (Abbot ID NOW)       <NA>         <NA>
#> 2                     <NA>                    Kaiser       <NA>         <NA>
#> 3                     <NA>                     Quest       <NA>         <NA>
#> 4                       AN                      <NA> 37.7619118 -122.4204929
#> 5                       AN                     Quest       <NA>         <NA>
#> 6                       AN                      <NA> 37.7815006 -122.4097896
#>                     geometry
#> 1 POINT (-122.4259 37.77001)
#> 2 POINT (-122.4641 37.78097)
#> 3 POINT (-122.4486 37.78699)
#> 4 POINT (-122.4205 37.76191)
#> 5 POINT (-122.4924 37.78198)
#> 6  POINT (-122.4098 37.7815)