San Francisco COVID-19 Testing Locations

A list of testing locations including address and coordinates for mapping. More information about the data available here

covid19sf_test_loc

Format

An object class sf and data.frame with 17 variables

id

Location ID

medical_home

Medical home

name

The medical name

address

The medical address

phone_number

The medical phone number

phone_number_formatted

The medical phone number formatted

testing_hours

The medical testing hours

popup_or_permanent

The medical testing type, c("Permanent", "Pop-Up" )

location_type

The medical location type, c("Private", "Public" )

eligibility

Eligibility information for accessing testing at this location

cta_text

The call to action used for the web map

cta_link

The call to action link for the button on the web map

sample_collection_method

The method for collecting samples at the lab

lap

The lab name

latitude

The medical latitude point

longitude

The medical longitude point

geometry

The medical geometry details

Source

San Francisco, Department of Public Health - Population Health Division through San Francisco Opne Data protal website.

Details

The dataset contains the San Francisco testing location information

Examples

data(covid19sf_test_loc)

head(covid19sf_test_loc)
#> Simple feature collection with 6 features and 16 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -122.4486 ymin: 37.7187 xmax: -122.3849 ymax: 37.78699
#> Geodetic CRS:  WGS 84
#>   id                                             medical_home
#> 1  6                                                     UCSF
#> 2  1                                            Carbon Health
#> 3 11                               Northeast Medical Services
#> 4 43 SF Department of Public Health in partnership with Color
#> 5 25 SF Department of Public Health in partnership with Color
#> 6 38                                                     <NA>
#>                              name                 address phone_number
#> 1           Laurel Heights Campus      3333 California St   4158857478
#> 2   Carbon Health - Castro Clinic          1998 Market St   4157926040
#> 3            San Bruno Ave Clinic      2574 San Bruno Ave   4153919686
#> 4                  Alice Griffith 2600 Arelious Walker Dr          311
#> 5           GLIDE Memorial Church        330 Ellis Street          311
#> 6 Test the People by renegade.bio            2730 21st St   5103273540
#>   phone_number_formatted                                  testing_hours
#> 1         (415) 885-7478                                    9am-5pm M-F
#> 2         (415) 792-6040                                        9am-7pm
#> 3         (415) 391-9686                                 8:30am-5pm M-F
#> 4             Dial 3-1-1           Saturday, 11/20 only from 10am - 3pm
#> 5             Dial 3-1-1                Tuesdays & Wednesdays, 10am-4pm
#> 6         (510) 327-3540 M, W, and F from 10 AM to 3PM, except holidays
#>   popup_or_permanent location_type
#> 1             Pop-Up       Private
#> 2          Permanent       Private
#> 3          Permanent       Private
#> 4             Pop-Up        Public
#> 5             Pop-Up        Public
#> 6             Pop-Up       Private
#>                                                         eligibility
#> 1                                              Must be UCSF patient
#> 2                                     Must be Carbon Health patient
#> 3 Must be NEMS patient. May register as a new NEMS patient for free
#> 4                                  Everyone. No insurance required.
#> 5                                  Everyone. No insurance required.
#> 6                                  Everyone. No insurance required.
#>                                 cta_text
#> 1            Physician referral required
#> 2                  Sign up with provider
#> 3                  Sign up with provider
#> 4                                Drop in
#> 5                               Drop in.
#> 6 Drop in or call ahead for appointments
#>                                                                                                                                                                                                                               cta_link
#> 1                                                                                                                                                                                                                     tel://4158857478
#> 2 https://patient.carbonhealth.com/#/schedule?practiceId=5bdaef44-8ff0-439f-99d7-3285afcc6911&vstId=0f02af4e-1d1c-4a4b-a0f2-e67fca98dd77&vstrId=f743d1d1-d2a1-4fce-a5f9-2e29339b2406&specialtyIds=652c64cc-9fcb-442f-b71e-266390ef2f63
#> 3                                                                                                                                                                                                        https://www.nems.org/covid19/
#> 4                                                                                                                                                                                                           https://sf.gov/GetTestedSF
#> 5                                                                                                                                                                                                           https://sf.gov/GetTestedSF
#> 6                                                                                                                                                                                                       https://www.testthepeople.org/
#>   sample_collection_method                       lab   latitude    longitude
#> 1                     <NA>                      UCSF       <NA>         <NA>
#> 2                     <NA> Rapid test (Abbot ID NOW)       <NA>         <NA>
#> 3                     <NA>                      <NA>       <NA>         <NA>
#> 4                     <NA>                      <NA> 37.7187018 -122.3849448
#> 5                     <NA>                      <NA> 37.7852341  -122.413658
#> 6                     <NA>                      <NA>  37.757629 -122.4094428
#>                     geometry
#> 1 POINT (-122.4486 37.78699)
#> 2 POINT (-122.4259 37.77001)
#> 3  POINT (-122.4042 37.7291)
#> 4  POINT (-122.3849 37.7187)
#> 5 POINT (-122.4137 37.78523)
#> 6 POINT (-122.4094 37.75763)