COVID-19 Vaccine Doses Given to San Franciscans by Demographics

This dataset represents doses of COVID-19 vaccine administered in California to residents of San Francisco. The data is broken down by multiple demographic slices. The three dose types are counted separately, i.e. (1) first doses administered as a part of a two-dose vaccination, (2) second doses administered as part of a two-dose vaccination, and (3) single-dose vaccines administered. here

NOTE: This dataset is no longer supported and will be deprecated on the next release (v0.1.3). The covid19sf_population dataset is an alternative for covid19sf_vaccine_demo.

covid19sf_vaccine_demo

Format

An object class data.frame with 15 variables

overall_segment

Segment (universe) of analysis. Unique combination of administering_provider_type, age_group, and demographic_group. Filter to a single option to derive meaningful totals.

administering_provider_type

Providers included in a given overall_segment. Two possible values: 'All' (including SF DPH) or 'DPH Only'

age_group

Age range included in a given overall_segment

demographic_group

Type of demographic group included in a given overall_segment (e.g. Age, Race/Ethnicity)

demographic_subgroup

Specific demographic group counted in a given record (e.g. 16-24, Asian)

demographic_subgroup_sort_order

Numeric sort order for all demographic_subgroups. Convenient for maintaining consistent ordering across multiple data visualizations.

total_1st_doses

Total number of first doses administered

total_2nd_doses

Total number of second doses administered

total_single_doses

Total number of single dose vaccines administered

total_recipients

Total number of unique vaccine recipients

total_series_completed

Total number of individuals fully vaccinated (those having received the second dose of a two-dose vaccine or one dose of a single-dose vaccine)

subgroup_population

2018 5-year American Community Survey population estimates for given DEMOGRAPHIC_SUBGROUP

age_group_population

2018 5-year American Community Survey population estimates for overall AGE_GROUP

data_as_of

Timestamp for last update date in source system

data_loaded_at

Timestamp when the record (row) was most recently updated in Socrata

Source

San Francisco, Department of Public Health - Population Health Division through San Francisco Opne Data protal website.

Details

The dataset contains a summary of COVID-19 vaccine doses given to San Franciscans by demographics

Examples

data(covid19sf_vaccine_demo)

head(covid19sf_vaccine_demo)
#>                                          overall_segment
#> 1 Ages 12+ by Age Bracket, Administered by All Providers
#> 2 Ages 12+ by Age Bracket, Administered by All Providers
#> 3 Ages 12+ by Age Bracket, Administered by All Providers
#> 4 Ages 12+ by Age Bracket, Administered by All Providers
#> 5      Ages 12+ by Age Bracket, Administered by DPH Only
#> 6 Ages 12+ by Age Bracket, Administered by All Providers
#>   administering_provider_type age_group demographic_group demographic_subgroup
#> 1               All Providers       12+       Age Bracket                25-34
#> 2               All Providers       12+       Age Bracket                12-17
#> 3               All Providers       12+       Age Bracket                65-74
#> 4               All Providers       12+       Age Bracket                18-24
#> 5                    DPH Only       12+       Age Bracket                12-17
#> 6               All Providers       12+       Age Bracket                35-44
#>   demographic_subgroup_sort_order total_1st_doses total_2nd_doses
#> 1                               5          150076          136628
#> 2                               3           36812           33979
#> 3                               9           75648           71846
#> 4                               4           51456           46027
#> 5                               3            5416            5256
#> 6                               6          119776          111380
#>   total_single_doses total_recipients total_series_completed
#> 1              14362           164830                 150990
#> 2                 18            36852                  33997
#> 3               3326            79023                  75172
#> 4               4297            55853                  50324
#> 5                  3             5690                   5259
#> 6              10914           130886                 122294
#>   subgroup_population age_group_population          data_as_of
#> 1              204639               791131 2021-12-15 06:00:11
#> 2               33938               791131 2021-12-15 06:00:11
#> 3               74120               791131 2021-12-15 06:00:11
#> 4               62127               791131 2021-12-15 06:00:11
#> 5               33938               791131 2021-12-15 06:00:11
#> 6              138390               791131 2021-12-15 06:00:11
#>        data_loaded_at
#> 1 2021-12-15 08:15:07
#> 2 2021-12-15 08:15:07
#> 3 2021-12-15 08:15:07
#> 4 2021-12-15 08:15:07
#> 5 2021-12-15 08:15:07
#> 6 2021-12-15 08:15:07