COVID-19 Vaccine Doses Given to San Franciscans by Demographics Over Time

This dataset represents doses of COVID-19 vaccine administered in California to San Francisco residents over time. The data is broken down by multiple demographic slices. The three dose types are counted separately, i.e. (1) first doses administered as a part of a two-dose vaccination, (2) second doses administered as part of a two-dose vaccination, and (3) single-dose vaccines administered. here

covid19sf_vaccine_demo_ts

Format

An object class data.frame with 19 variables

date_administered

Date vaccination administered

overall_segment

Segment (universe) of analysis. Unique combination of administering_provider_type, age_group, and demographic_group. Filter to a single option to derive meaningful totals.

administering_provider_type

Providers included in a given overall_segment. Two possible values: 'All' (including SF DPH) or 'DPH Only'

age_group

Age range included in a given overall_segment

demographic_group

Type of demographic group included in a given overall_segment (e.g. Age, Race/Ethnicity)

demographic_subgroup

Specific demographic group counted in a given record (e.g. 16-24, Asian)

demographic_subgroup_sort_order

Numeric sort order for all demographic_subgroup. Convenient for maintaining consistent ordering across multiple data visualizations.

new_1st_doses

Count of 1st doses administered for vaccines that take two doses to complete

new_2nd_doses

Count of 2nd doses administered for vaccines that take two doses to complete

new_single_doses

Count of doses administered for vaccines that take one dose to complete

new_series_completed

Count of individuals newly fully vaccinated on a given day (given the 2nd dose of a two-dose vaccine or one dose of a single dose vaccine)

new_recipients

Count of individuals vaccinated (with any dose) for the first time according to CA's records

cumulative_1st_doses

Cumulative total of 1st doses administered for vaccines that take two doses to complete

cumulative_2nd_doses

Cumulative total of 2nd doses administered for vaccines that take two doses to complete

cumulative_single_doses

Cumulative total of doses administered for vaccines that take one dose to complete

cumulative_series_completed

Cumulative total individuals fully vaccinated (given the 2nd dose of a two-dose vaccine or one dose of a single dose vaccine)

cumulative_recipients

Cumulative total individuals vaccinated (with any dose) according to CA's records

subgroup_population

American Community Survey population estimates for given demographic_subgroup

age_group_population

American Community Survey population estimates for overall age_group

Source

San Francisco, Department of Public Health - Population Health Division through San Francisco Opne Data protal website.

Details

The dataset contains a time series of COVID-19 vaccine doses given to San Franciscans by demographics

Examples

data(covid19sf_vaccine_demo_ts)

head(covid19sf_vaccine_demo_ts)
#>   date_administered                                           overall_segment
#> 1        2021-08-12 Ages 65+ by Race/Ethnicity, Administered by All Providers
#> 2        2021-08-13 Ages 65+ by Race/Ethnicity, Administered by All Providers
#> 3        2021-08-14 Ages 65+ by Race/Ethnicity, Administered by All Providers
#> 4        2021-08-15 Ages 65+ by Race/Ethnicity, Administered by All Providers
#> 5        2021-08-16 Ages 65+ by Race/Ethnicity, Administered by All Providers
#> 6        2021-08-17 Ages 65+ by Race/Ethnicity, Administered by All Providers
#>   administering_provider_type age_group demographic_group demographic_subgroup
#> 1               All Providers       65+    Race/Ethnicity              Unknown
#> 2               All Providers       65+    Race/Ethnicity              Unknown
#> 3               All Providers       65+    Race/Ethnicity              Unknown
#> 4               All Providers       65+    Race/Ethnicity              Unknown
#> 5               All Providers       65+    Race/Ethnicity              Unknown
#> 6               All Providers       65+    Race/Ethnicity              Unknown
#>   demographic_subgroup_sort_order new_1st_doses new_2nd_doses new_single_doses
#> 1                               9             0             0                1
#> 2                               9             1             0                2
#> 3                               9             0             0                0
#> 4                               9             0             0                0
#> 5                               9             0             0                3
#> 6                               9             1             1                0
#>   new_series_completed new_recipients cumulative_1st_doses cumulative_2nd_doses
#> 1                    1              1                  999                  774
#> 2                    2              3                 1000                  774
#> 3                    0              0                 1000                  774
#> 4                    0              0                 1000                  774
#> 5                    3              3                 1000                  774
#> 6                    1              1                 1001                  775
#>   cumulative_single_doses cumulative_series_completed cumulative_recipients
#> 1                     126                         900                  1126
#> 2                     128                         902                  1129
#> 3                     128                         902                  1129
#> 4                     128                         902                  1129
#> 5                     131                         905                  1132
#> 6                     131                         906                  1133
#>   subgroup_population age_group_population
#> 1                   0               135027
#> 2                   0               135027
#> 3                   0               135027
#> 4                   0               135027
#> 5                   0               135027
#> 6                   0               135027